+++ src/components/community/post/index.ts
@@ -1,9 +1,6 @@
-export * from './Complete';
 export * from './SectionInfo';
 export * from './Title';
-export * from './TitleImageDescription';
 export * from './Write';
-export * from './TypeCategory';
 export * from './CategorySelector';
 export * from './DraftModal';
 export * from './Description';
